# app/blueprints/docs/static/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "GCM Ops API",
    "version": "1.0.0",
    "description": "APIs internas para presentismo (horas trabajadas/extras, inasistencias), reportes, autenticación y administración de usuarios/desvinculaciones."
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local"
    },
    {
      "url": "https://api.tu-dominio.com",
      "description": "Producción"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      }
    },
    "schemas": {
...
      "DesvCreate": {
        "type": "object",
        "properties": {
          "dni": {
            "type": "string"
          },
          "nombre": {
            "type": "string"
          },
          "cuenta_area": {
            "type": "string"
          },
          "recinto": {
            "type": "string"
          },
          "fecha_ingreso": {
            "type": "string",
            "example": "2024-05-10"
          },
          "fecha_retiro": {
            "type": "string",
            "example": "2025-09-30"
          },
          "motivo": {
            "type": "string"
          },
          "tipo_contrato": {
            "type": "string"
          },
          "observaciones": {
            "type": "string"
          }
        },
        "required": [
          "dni",
          "nombre",
          "cuenta_area",
          "recinto",
          "fecha_retiro",
          "motivo"
        ]
      }
    },
...
  },
  "paths": {
...
    "/desvinculaciones/": {
      "get": {
        "tags": [
          "Desvinculaciones"
        ],
        "summary": "Listado (HTML) con filtros y paginación",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "per",
            "schema": {
              "type": "integer",
              "default": 100
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "desde",
            "schema": {
              "type": "string",
              "example": "2025-10-01"
            }
          },
          {
            "in": "query",
            "name": "hasta",
            "schema": {
              "type": "string",
              "example": "2025-10-31"
            }
          },
          {
            "in": "query",
            "name": "area",
            "schema": {
              "type": "string",
              "example": "BAT"
            }
          },
          {
            "in": "query",
            "name": "recinto",
            "schema": {
              "type": "string",
              "example": "Noviciado"
            }
          },
          {
            "in": "query",
            "name": "q",
            "schema": {
              "type": "string",
              "example": "perez"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "HTML"
          }
        }
      }
    },
    "/desvinculaciones/export": {
      "get": {
        "tags": [
          "Desvinculaciones"
        ],
        "summary": "Exportar desvinculaciones (XLSX/CSV)",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "desde",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "hasta",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "area",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "recinto",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "format",
            "schema": {
              "type": "string",
              "enum": [
                "xlsx",
                "csv"
              ],
              "default": "xlsx"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Archivo XLSX/CSV",
            "content": {
              "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": {},
              "text/csv": {}
            }
          }
        }
      }
    },
    "/desvinculaciones/new": {
      "get": {
        "tags": [
          "Desvinculaciones"
        ],
        "summary": "Formulario de creación (HTML)",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "HTML"
          }
        }
      },
      "post": {
        "tags": [
          "Desvinculaciones"
        ],
        "summary": "Crear desvinculación",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/DesvCreate"
              }
            }
          }
        },
        "responses": {
          "302": {
            "description": "Redirige si OK"
          },
          "200": {
            "description": "HTML con errores"
          }
        }
      }
    },
    "/desvinculaciones/{id}/edit": {
      "get": {
        "tags": [
          "Desvinculaciones"
        ],
        "summary": "Formulario de edición (HTML)",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "HTML"
          },
          "404": {
            "description": "No encontrado"
          }
        }
      },
      "post": {
        "tags": [
          "Desvinculaciones"
        ],
        "summary": "Actualizar desvinculación",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/DesvCreate"
              }
            }
          }
        },
        "responses": {
          "302": {
            "description": "Redirige si OK"
          },
          "200": {
            "description": "HTML con errores"
          },
          "404": {
            "description": "No encontrado"
          }
        }
      }
    },
    "/desvinculaciones/{id}/delete": {
      "post": {
        "tags": [
          "Desvinculaciones"
        ],
        "summary": "Eliminar desvinculación",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Redirige al listado"
          },
          "404": {
            "description": "No encontrado"
          }
        }
      }
    },
    "/desvinculaciones/cargar": {
      "get": {
        "tags": [
          "Desvinculaciones"
        ],
        "summary": "Pantalla de carga/gestión (HTML)",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "HTML"
          }
        }
      }
    },
    "/desvinculaciones/carga-masiva": {
      "get": {
        "tags": [
          "Desvinculaciones"
        ],
        "summary": "Formulario de carga masiva (HTML)",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "HTML"
          }
        }
      },
      "post": {
        "tags": [
          "Desvinculaciones"
        ],
        "summary": "Procesar carga masiva (CSV)",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": [
                  "file"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Resultado (HTML/flash)"
          },
          "400": {
            "description": "Archivo inválido"
          }
        }
      }
    },
    "/desvinculaciones/carga-masiva/plantilla": {
      "get": {
        "tags": [
          "Desvinculaciones"
        ],
        "summary": "Descargar plantilla CSV",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Archivo CSV",
            "content": {
              "text/csv": {}
            }
          }
        }
      }
    },
...
  }
}